Module Css_gen.Expert

val should_validate : bool Core.ref

By default, the Css_gen constructors validate that all values are well-formed. This can be useful for debugging, as it'll throw an exception when the programmer makes a mistake. However, it also incurs a heavy cost and should be avoided in tight loops. should_validate allows the programmer to disable css value validation when necessary.