Module Ui_effect.Expert

val handle : unit t -> unit

handle t looks up the Handler.handle function in the table of Defined functions, unwraps the Event.t back into its underlying Action.t, and applies the two. This is only intended for internal use by this library, specifically by the attribute code.

type hide =

| T : ('a t * ( 'a -> unit )) -> hide

val handlers : ( hide -> unit ) Core.Hashtbl.M(Core.Int).t

val of_fun : f:( callback:( 'a -> unit ) -> unit ) -> 'a t