Module Ui_effect.For_testing

module Svar : sig ... end

You can think of an Svar.t as like an Ivar.t whose purpose is to allow us to implement of_svar below.

val of_svar_fun : ( 'query -> 'result Svar.t ) -> 'query -> 'result t

Create an effect from a function that returns an Svar.t. This is mostly useful in testing, to emulate a 'query -> 'result Deferred.t function that does not return immediately. You may find Query_response_tracker a more convenient interface than using of_svar directly.

module Query_response_tracker : sig ... end

val of_query_response_tracker : ( 'query, 'result ) Query_response_tracker.t -> 'query -> 'result t