Module Vdom_input_widgets

module type Display = sig ... end

module type Equal = sig ... end

module type Enum = sig ... end

module type Set = sig ... end

module type Enum_set = sig ... end

module Decimal : Core.Stringable.S with type t = float

Decimal can be used to create decimal number inputs

module Validated : sig ... end

module Dropdown : sig ... end

module Checkbox : sig ... end

module Checklist : sig ... end

module Multi_select : sig ... end

Creates a multiple-selection list. This is different from Checklist in that the <select multiple> element is used underneath, rather than independent checkboxes. As a result:

module Entry : sig ... end

module Button : sig ... end

module Radio_buttons : sig ... end

module File_select : sig ... end

Elements for selecting files from local disk