Module Vdom_input_widgets.Dropdown

val of_values : ?extra_attrs:Virtual_dom.Vdom.Attr.t list -> ?disabled:bool -> (module Equal with type t = 'a) -> 'a list -> selected:'a -> on_change:( 'a -> unit Virtual_dom.Vdom.Effect.t ) -> Virtual_dom.Vdom.Node.t

Creates a dropdown that automatically updates when the current value changes and emits typed actions when the user selects a different item.

val of_values_opt : ?extra_attrs:Virtual_dom.Vdom.Attr.t list -> ?disabled:bool -> (module Equal with type t = 'a) -> 'a list -> selected:'a option -> on_change:( 'a option -> unit Virtual_dom.Vdom.Effect.t ) -> Virtual_dom.Vdom.Node.t

Same as of_values, but includes a blank first entry to represent None.

val of_enum : ?extra_attrs:Virtual_dom.Vdom.Attr.t list -> ?disabled:bool -> (module Enum with type t = 'a) -> selected:'a -> on_change:( 'a -> unit Virtual_dom.Vdom.Effect.t ) -> Virtual_dom.Vdom.Node.t

Instead of passing individual values, uses all from the module to determine the items and order.

val of_enum_opt : ?extra_attrs:Virtual_dom.Vdom.Attr.t list -> ?disabled:bool -> (module Enum with type t = 'a) -> selected:'a option -> on_change:( 'a option -> unit Virtual_dom.Vdom.Effect.t ) -> Virtual_dom.Vdom.Node.t

Instead of passing individual values, uses all from the module to determine the items and order. Include a blank entry to represent None