Module Vdom_input_widgets.Multi_select

Creates a multiple-selection list. This is different from Checklist in that the <select multiple> element is used underneath, rather than independent checkboxes. As a result:

Clicking an element makes it the only one selected.

Ctrl+clicking an element toggles it independently.

Shift+clicking an element selects everything between it and the previous click.

Clicking and dragging selects a range.

module Repeated_click_behavior : sig ... end

val of_values : ?extra_attrs:Virtual_dom.Vdom.Attr.t list -> ?repeated_click_behavior:Repeated_click_behavior.t -> ?disabled:bool -> ?size:int -> (module Set with type comparator_witness = 'cmp and type t = 'a) -> 'a list -> selected:( 'a, 'cmp ) Core.Set.t -> on_change:( ( 'a, 'cmp ) Core.Set.t -> unit Virtual_dom.Vdom.Effect.t ) -> Virtual_dom.Vdom.Node.t

val of_enum : ?extra_attrs:Virtual_dom.Vdom.Attr.t list -> ?repeated_click_behavior:Repeated_click_behavior.t -> ?disabled:bool -> ?size:int -> (module Enum_set with type comparator_witness = 'cmp and type t = 'a) -> selected:( 'a, 'cmp ) Core.Set.t -> on_change:( ( 'a, 'cmp ) Core.Set.t -> unit Virtual_dom.Vdom.Effect.t ) -> Virtual_dom.Vdom.Node.t