Module Vdom_input_widgets.Radio_buttons

module Style : sig ... end

val of_values : ?extra_attrs:Virtual_dom.Vdom.Attr.t list -> ?disabled:bool -> ?style:Style.t -> (module Equal with type t = 'a) -> name:string -> on_click:( 'a -> unit Virtual_dom.Vdom.Effect.t ) -> selected:'a option -> 'a list -> Virtual_dom.Vdom.Node.t

Creates a radio button input with a button for each of the corresponding values passed in. Extra attrs for individual buttons can be passed through extra_element_attrs, which gets access to whether the current element is selected or not.

val of_values_horizontal : ?extra_attrs:Virtual_dom.Vdom.Attr.t list -> ?disabled:bool -> ?style:Style.t -> (module Equal with type t = 'a) -> name:string -> on_click:( 'a -> unit Virtual_dom.Vdom.Effect.t ) -> selected:'a option -> 'a list -> Virtual_dom.Vdom.Node.t

Identical to of_values, but adds some style tags so that the list is displayed horizontally instead of vertically.