Module Vdom_input_widgets.Validated

type 'a t

include Ppx_compare_lib.Equal.S1 with type 'a t := 'a t

val equal : ( 'a -> 'a -> bool ) -> 'a t -> 'a t -> bool

include Sexplib0.Sexpable.S1 with type 'a t := 'a t

val t_of_sexp : ( Sexplib0__.Sexp.t -> 'a ) -> Sexplib0__.Sexp.t -> 'a t

val sexp_of_t : ( 'a -> Sexplib0__.Sexp.t ) -> 'a t -> Sexplib0__.Sexp.t

include Bin_prot.Binable.S1 with type 'a t := 'a t

val bin_shape_t : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_size_t : ( 'a, 'a t ) Bin_prot.Size.sizer1

val bin_write_t : ( 'a, 'a t ) Bin_prot.Write.writer1

val bin_read_t : ( 'a, 'a t ) Bin_prot.Read.reader1

val __bin_read_t__ : ( 'a, int -> 'a t ) Bin_prot.Read.reader1

val bin_writer_t : ( 'a, 'a t ) Bin_prot.Type_class.S1.writer

val bin_reader_t : ( 'a, 'a t ) Bin_prot.Type_class.S1.reader

val bin_t : ( 'a, 'a t ) Bin_prot.Type_class.S1.t

include Ppx_compare_lib.Comparable.S1 with type 'a t := 'a t

val compare : ( 'a -> 'a -> int ) -> 'a t -> 'a t -> int

type 'a update

The update type is provided when a user changes the input in an entry. It cannot be used on its own, but it can be applied to an existing t to access the new current or last values.

val equal_update : ( 'a -> 'a -> bool ) -> 'a update -> 'a update -> bool

val sexp_of_update : ( 'a -> Sexplib0.Sexp.t ) -> 'a update -> Sexplib0.Sexp.t

val update_of_sexp : ( Sexplib0.Sexp.t -> 'a ) -> Sexplib0.Sexp.t -> 'a update

val bin_shape_update : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_size_update : 'a Bin_prot.Size.sizer -> 'a update Bin_prot.Size.sizer

val bin_write_update : 'a Bin_prot.Write.writer -> 'a update Bin_prot.Write.writer

val bin_writer_update : 'a Bin_prot.Type_class.writer -> 'a update Bin_prot.Type_class.writer

val bin_read_update : 'a Bin_prot.Read.reader -> 'a update Bin_prot.Read.reader

val __bin_read_update__ : 'a Bin_prot.Read.reader -> ( int -> 'a update ) Bin_prot.Read.reader

val bin_reader_update : 'a Bin_prot.Type_class.reader -> 'a update Bin_prot.Type_class.reader

val bin_update : 'a Bin_prot.Type_class.t -> 'a update Bin_prot.Type_class.t

val compare_update : ( 'a -> 'a -> int ) -> 'a update -> 'a update -> int

val initial_empty : 'a t

initial_empty is a special value that, when provided to an Entry.validate widget, will reset the contents regardless of whether the user has focus or not.

val is_initial_empty : _ t -> bool

val return : 'a -> 'a t

val get_current : 'a t -> 'a option

val get_last : 'a t -> 'a option

val get_error : 'a t -> string option

val update : 'a t -> 'a update -> 'a t

val lift : (module Core.Stringable with type t = 'a) -> (module Core.Stringable with type t = 'a t)