Module Vdom_keyboard

module Grouped_help_text : sig ... end

A Grouped_help_text.t is similar to a Help_text.t, but allows the user to organize the commands into groups.

module Help_text : sig ... end

A Help_text.t represents the documentation for a collection of commands. It can be displayed as a Vdom node.

module Keyboard_event_handler : sig ... end

A Keyboard_event_handler.t is a collections of actions (commands and disabled keys) that can be used to handle keyboard events and to produce help text.

module Keystroke : sig ... end

module Variable_keyboard_event_handler : sig ... end

A Variable_keyboard_event_handler.t provides a way of representing a keyboard event handler that contains both a (possibly empty) constant set of actions and a variable set of actions that depends on some 'env variable. In order to handle keyboard events or produce help text, it must first be converted to a Keyboard_event_handler.t using the function to_const_handler.

module Keyboard_event : sig ... end

val with_keyboard_handler : Virtual_dom.Vdom.Node.t -> Keyboard_event_handler.t -> Virtual_dom.Vdom.Node.t