Module Grouped_help_text.View_spec

View_spec is almost identical to Help_text.View_spec, but additionally allows the user to customize how to display group names.

type t = {

core_spec : Help_text.View_spec.t;

group_name : Group_name.t -> Virtual_dom.Vdom.Node.t;

}

val plain : t

val with_classes : group_name_class:string -> key_class:string -> plain_text_class:string -> t

with_classes behaves the same as Help_text.View_spec.with_classes as far as the core_spec, and additionally converts group names to text nodes and wraps them in divs with the given group name class.