Module Vdom_keyboard.Help_text

A Help_text.t represents the documentation for a collection of commands. It can be displayed as a Vdom node.

This can be used to create a web ui help menu.

module View_spec : sig ... end

module Command : sig ... end

A Command.t represents the help text for a single command.

type t

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t -> t -> int

val empty : t

val is_empty : t -> bool

val of_command_list : Command.t list -> t

val add_command : t -> Command.t -> t

add_command is linear in the number of commands in t.

val commands : t -> Command.t list

val view : ?sep:string -> t -> View_spec.t -> Virtual_dom.Vdom.Node.t

view displays a help text table with one row per command.

view_rows is similar to view, but returns a list of row nodes instead of wrapping them in a table node.

val view_rows : ?sep:string -> t -> View_spec.t -> Virtual_dom.Vdom.Node.t list