Module Help_text.Command

A Command.t represents the help text for a single command.

type t = {

keys : Keystroke.t list;

description : string;

}

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

module Format : sig ... end

val view : t -> View_spec.t -> Format.t -> Virtual_dom.Vdom.Node.t

view displays a help text line for a single command. For instance, this can be used to display the help text for opening a help menu.