Module Keyboard_event_handler.Action

type t =

| Command of Command.t	(*
A disabled key is essentially a command that prevents the default handler of that key (by returning Effect.Prevent_default), and does nothing else. Users can choose to omit disabled keys from the help menu.
*)

| Disabled_key of Keystroke.t

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

val command : Command.t -> t

val disabled_key : Keystroke.t -> t

val is_command : t -> bool

val is_disabled_key : t -> bool

val command_val : t -> Command.t option

val disabled_key_val : t -> Keystroke.t option

module Variants : sig ... end

val get_help_text : t -> Help_text.Command.t