Module Variable_keyboard_event_handler.Variable_handler_command

Variable_handler_command and Variable_handler_action provide a way of representing commands whose keys, description and group are constant, but whose handler varies with the 'env variable.

This is useful because it provides a way of generating help text for these commands that does not depend on or vary with the 'env variable.

type 'env t = {

keys : Keystroke.t list;

description : string;

group : Grouped_help_text.Group_name.t option;

handler : 'env -> Keyboard_event_handler.Handler.t;

}

val get_help_text : _ t -> Help_text.Command.t