Module Attr.Global_listeners

Hooks to set mouse events listeners on window. This is needed as if we only set them on individual elements we will miss ones that happen outside of the viewport

https://coderwall.com/p/79hkbw/js-mouse-events-that-work-even-when-mouse-is-moved-outside-the-window

val mouseup : ( Js_of_ocaml.Dom_html.mouseEvent Js_of_ocaml.Js.t -> unit Ui_effect.t ) -> Virtual_dom__Attr.t

val mousemove : ( Js_of_ocaml.Dom_html.mouseEvent Js_of_ocaml.Js.t -> unit Ui_effect.t ) -> Virtual_dom__Attr.t

module For_testing : sig ... end