Module Hooks.Make

Parameters

module S : S

Signature

val create : S.Input.t -> t

name is a unique identifier that is treated like the names of regular attributes like "id" and "class" in <div id=... class=...> in that there can only be one attribute with the same name on an element, and that hooks are diffed only if the same hook has the same name between stabilizations.

module For_testing : sig ... end

The type-id provided here can be used to pull out the input value for an instance of this hook for testing-purposes.