Module Attr.Unmerged_warning_mode

type t =

| No_warnings

| All_warnings

| Stop_after_quota of Base.int

Controls whether to_raw should print warning messages when one attribute overrides another of the same name (for example, if there are two title attributes, the second will end up in the result, and a warning will be emitted).

If an application never emits any warnings, it is probably safe to always use good merge semantics (that is, use many or @ to combine lists of attributes) for everything.

val current : t Base.ref

Defaults to Stop_after_quota 100

module For_testing : sig ... end