Module Effect.Expert

val handle : Js_of_ocaml.Dom_html.event Js_of_ocaml.Js.t -> unit t -> unit

handle t looks up the Handler.handle function in the table of Defined functions, unwraps the Effect.t back into its underlying Action.t, and applies the two. This is only intended for internal use by this library, specifically by the attribute code.

val handle_non_dom_event_exn : unit t -> unit

handle_non_dom_event_exn is the same as handle except that it raises in any case that would have required the #Dom_html.event Js.t. In particular, this can be to feed Actions back to the system that are not triggered by events from the DOM and do not have a corresponding #Dom_html.event Js.t.