Module Virtual_dom_test_helpers.Node_helpers

type element = {

tag_name : string;

attributes : (string * string) list;

string_properties : (string * string) list;

bool_properties : (string * bool) list;

styles : (string * string) list;

handlers : (string * Handler.t) list;

hooks : (string * Virtual_dom.Vdom.Attr.Hooks.For_testing.Extra.t) list;

key : string option;

children : t list;

}

and t =

| Text of string

| Element of element

| Widget of Core.Sexp.t

Roughly analogous to Vdom.Node.t, but more easily inspectable and represented as a pure OCaml type.

val sexp_of_element : element -> Sexplib0.Sexp.t

val sexp_of_t : t -> Sexplib0.Sexp.t

val map : t -> f:( t -> [ `Continue | `Replace_with of t ] ) -> t

val is_tag : tag:string -> t -> bool

val has_class : cls:string -> t -> bool

val select : t -> selector:string -> t list

val select_first : t -> selector:string -> t option

val select_first_exn : t -> selector:string -> t

val to_string_html : ?filter_printed_attributes:( string -> bool ) -> t -> string

val inner_text : t -> string

val unsafe_convert_exn : Virtual_dom.Vdom.Node.t -> t

val trigger : ?extra_fields:(string * Js_of_ocaml.Js.Unsafe.any) list -> t -> event_name:string -> unit

val trigger_hook : t -> type_id:( 'a -> unit Virtual_dom.Vdom.Effect.t ) Core.Type_equal.Id.t -> name:string -> arg:'a -> unit

When a hook-based attribute build from an event-returning function, this function will find the hook, extract the value, call that function with arg, and schedule the resulting function.

val get_hook_value : t -> type_id:'a Core.Type_equal.Id.t -> name:string -> 'a

Given an element, this function attempts to retrieve a hook with the name name, and the type-id from the hooks For_testing module.

module User_actions : sig ... end

Convenience functions for trigger, closely modeling user interactions.